www-1.0/requirements.txt
Django==2.2.2
dj-database-url==0.5.0
django-bootstrap4==0.0.8
django-crispy-forms==1.7.2
Pillow
psycopg2
python-decouple==3.1
sqlparse==0.3.0
django-formtools
pandas
numpy
plotly
xgboost==0.81
matplotlib
dash==1.3.1
dash-core-components==1.2.1
dash-html-components==1.0.1
dash-renderer==1.1.0
xlrd==1.2.0
